+++ lyrebird/mock/blueprints/apis/snapshot_import.py
@@ -1,5 +1,5 @@
 from flask_restful import Resource, request
-from flask import redirect, Response
+from flask import redirect
 from lyrebird import application
 from lyrebird.mock import context
 from lyrebird import log
@@ -17,6 +17,7 @@ from ....version import VERSION
 logger = log.get_logger()
 
 importSnapshotUrl = None
+
 
 class SanpshotImport(Resource):
     """
@@ -37,7 +38,7 @@ class SanpshotImport(Resource):
             return application.make_fail_response(msg="has no param : parentNode.id")
         if importSnapshotUrl == None:
             return application.make_fail_response(msg="has no param : importSnapshotUrl")
-        
+
         # get config
         conf = application.config.raw()
         logger.debug(f'conf:\n {conf}')
@@ -71,6 +72,7 @@ class SanpshotImport(Resource):
 
         # load lyrebird_prop
         decompressed_innermost_path_list = []
+
         def find_prop(path):
             file_list = os.listdir(path)
             for file in file_list:
@@ -94,7 +96,6 @@ class SanpshotImport(Resource):
             snapshot_prop_obj=mockdata_node
         )
 
-        
         for file in os.listdir(mock_data_innermost_path):
             if "." not in file:
                 file_path = os.path.join(mock_data_innermost_path, file)
@@ -105,54 +106,3 @@ class SanpshotImport(Resource):
                 )
         context.application.data_manager.activate(_id=mockdata_node["id"])
         return application.make_ok_response()
-
-
-class SnapshotExport(Resource):
-    def post(self):
-        if not request.json.get("nodeId"):
-            return application.make_fail_response(f"获取导出nodeId失败")
-        node_id = request.json.get("nodeId")
-
-        # get config
-        conf = application.config.raw()
-        logger.debug(f'conf:\n {conf}')
-        mock_data_repositories = conf.get("mock.data")
-        if not conf.get("snapshot"):
-            return application.make_fail_response(msg=" config snapshot not defined")
-        snapshot_export_repositories = conf.get("snapshot")
-        logger.debug(f'snapshot_repositories: {snapshot_export_repositories}')
-        if not os.path.exists(snapshot_export_repositories):
-            os.makedirs(snapshot_export_repositories)
-            logger.debug(f"make dir {snapshot_export_repositories} success")
-
-        # defind output dir
-        export_dir_name  = str(time.strftime(
-            "%Y-%m-%d-%H-%M-%S", time.localtime()))+"-" + str(uuid.uuid4())
-
-        # export dm
-        context.application.data_manager.export_snapshot(
-            node_id=node_id,
-            export_root_path=snapshot_export_repositories,
-            export_dir_name=export_dir_name,
-            mock_data_repositories=mock_data_repositories
-            )
-        
-        # tar export dir
-        t = tarfile.open(f'{snapshot_export_repositories}/{export_dir_name}.gz', "w:gz")
-        for root, dir, files in os.walk(f'{snapshot_export_repositories}/{export_dir_name}'):
-            for file in files:
-                fullpath = os.path.join(root, file)
-                t.add(fullpath)
-        t.close()
-
-        # stream return
-        def generate():
-            with open(f'{snapshot_export_repositories}/{export_dir_name}.gz', "rb") as f:
-                while True:
-                    data = f.read(2048)  # 每次读取指定的长度
-                    import time
-                    time.sleep(1)
-                    if not data:
-                        break
-                    yield data
-        return Response(generate())
